+++ .github/workflows/tag-release.yml
@@ -1,0 +1,26 @@
+name: Tag Release
+
+on:
+  push:
+    branches: [main]
+    paths:
+      - version.go
+
+permissions:
+  contents: write
+
+jobs:
+  tag:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6.0.2
+        with:
+          fetch-tags: true # so a duplicate tag makes `git tag` fail
+
+      - name: Tag the version from version.go
+        run: |
+          version="$(grep -oE 'Version = "[0-9]+\.[0-9]+\.[0-9]+"' version.go | grep -oE '[0-9]+\.[0-9]+\.[0-9]+')"
+          echo "Tagging v$version"
+          git tag "v$version"
+          git push origin "v$version"
